feat(InputParser.kt): 로또 번호 자료형을 Long -> Int로 변경
Class Lotto에서 numbers를 List<Int>로 받는 것을 확인하고, 이에 맞춰 수정함

+++ src/main/kotlin/lotto/InputParser.kt
@@ -1,16 +1,16 @@
 package lotto
 
 object InputParser {
-    fun parsePaymentInput(input: String): Long {
-        return input.toLong()
+    fun parsePaymentInput(input: String): Int {
+        return input.toInt()
     }
 
-    fun parseWinningNumbersInput(input: String): Iterable<Long> {
-        return input.split(",").map { it -> it.toLong() }
+    fun parseWinningNumbersInput(input: String): Iterable<Int> {
+        return input.split(",").map { it -> it.toInt() }
     }
 
-    fun parseBonusNumberInput(input: String): Long {
-        return input.toLong()
+    fun parseBonusNumberInput(input: String): Int {
+        return input.toInt()
     }
 
 }
